Move DeleteTrackResult to media.tracks package
