+++ core-ui/src/main/java/fr/nihilus/music/core/ui/actions/DeleteTracksAction.kt
@@ -17,7 +17,7 @@
 package fr.nihilus.music.core.ui.actions
 
 import fr.nihilus.music.core.media.MediaId
-import fr.nihilus.music.media.provider.DeleteTracksResult
+import fr.nihilus.music.media.tracks.DeleteTracksResult
 import fr.nihilus.music.media.tracks.TrackRepository
 import javax.inject.Inject
