+++ media/src/main/java/fr/nihilus/music/media/tracks/DeleteTracksResult.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package fr.nihilus.music.media.provider
+package fr.nihilus.music.media.tracks
 
 import android.app.PendingIntent
