+++ media/src/test/java/fr/nihilus/music/media/provider/MediaStoreDaoTest.kt
@@ -31,6 +31,7 @@ import fr.nihilus.music.core.test.coroutines.CoroutineTestRule
 import fr.nihilus.music.core.test.failAssumption
 import fr.nihilus.music.media.os.MediaStoreDatabase
 import fr.nihilus.music.media.os.SimulatedFileSystem
+import fr.nihilus.music.media.tracks.DeleteTracksResult
 import io.kotest.assertions.withClue
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldContainExactly
